+++ lib/campaign/CampaignScenarioPrologEpilog.h
@@ -9,6 +9,8 @@
  */
 #pragma once
 
+VCMI_LIB_NAMESPACE_BEGIN
+
 struct DLL_LINKAGE CampaignScenarioPrologEpilog
 {
 	bool hasPrologEpilog = false;
@@ -24,3 +26,5 @@ struct DLL_LINKAGE CampaignScenarioPrologEpilog
 		h & prologText;
 	}
 };
+
+VCMI_LIB_NAMESPACE_END
