+++ lib/gameState/CGameState.h
@@ -29,39 +29,10 @@ struct EventCondition;
 struct CampaignTravel;
 class CStackInstance;
 class CGameStateCampaign;
+struct SThievesGuildInfo;
 
 template<typename T> class CApplier;
 class CBaseForGSApply;
-
-struct DLL_LINKAGE SThievesGuildInfo
-{
-	std::vector<PlayerColor> playerColors; //colors of players that are in-game
-
-	std::vector< std::vector< PlayerColor > > numOfTowns, numOfHeroes, gold, woodOre, mercSulfCrystGems, obelisks, artifacts, army, income; // [place] -> [colours of players]
-
-	std::map<PlayerColor, InfoAboutHero> colorToBestHero; //maps player's color to his best heros'
-
-    std::map<PlayerColor, EAiTactic::EAiTactic> personality; // color to personality // ai tactic
-	std::map<PlayerColor, si32> bestCreature; // color to ID // id or -1 if not known
-
-//	template <typename Handler> void serialize(Handler &h, const int version)
-//	{
-//		h & playerColors;
-//		h & numOfTowns;
-//		h & numOfHeroes;
-//		h & gold;
-//		h & woodOre;
-//		h & mercSulfCrystGems;
-//		h & obelisks;
-//		h & artifacts;
-//		h & army;
-//		h & income;
-//		h & colorToBestHero;
-//		h & personality;
-//		h & bestCreature;
-//	}
-
-};
 
 struct DLL_LINKAGE RumorState
 {
